web/satellite: fix mobile header not visible on ios safari and chrome
This change fixes an issue where the navigation header will not show on iOS (safari and chrome).
This is apparently a known issue where the Webkit viewport height is taller than the visible viewport.
More information here; https://bugs.webkit.org/show_bug.cgi?id=141832 , https://stackoverflow.com/a/56671096

This change applies the fix on all iOS browsers (using Webkit) except for Firefox, in which this fix introduces a strange behaviour
when a virtual keyboard is opened and dismissed, while not solving the issue because it doesn't exist here.

Issue: https://github.com/storj/storj/issues/5153

Change-Id: I19fc8018012bb70a5eda3eb21d44bfeedd08411c